Integrate graph plot and statistic calculation
To better understand the performance of iterative method and fast doubling method, we conducted data analysis using statistical methods. We calculated the average run times of 100 iterations and removed outliers beyond the 95% confidence interval to obtain more reliable data.

To visualize the data more conveniently, we used the matplotlib library to create graphs.

To optimize memory usage, we release the unuse memories by 'kfree' function in the 'string_iterative' function.